requirements.txt
black
cookiecutter
isort
pytest
pytest-cookies
wheel
ruamel.yaml
jsonschema
requests
